+++ src/Microsoft.ApplicationInsights.AspNetCore/TelemetryInitializers/WebSessionTelemetryInitializer.cs
@@ -9,6 +9,7 @@
     {
         private const string WebSessionCookieName = "ai_session";
 
+        /// <inheritdoc />
         public WebSessionTelemetryInitializer(IHttpContextAccessor httpContextAccessor)
              : base(httpContextAccessor)
         {